pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-template"
version = "0.1.0"
description = "Whatsapp Agent Template"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain_community>=0.3.17,<0.4",
    "langchain_openai>=0.3.10,<0.4",
    "langchain-google-genai>=2.1.2,<3",
    "langgraph>=0.3.20,<0.4",
    "langgraph-sdk>=0.1.61, <1",
    "langchain-mcp-adapters>=0.0.5,<1",
    "langgraph_supervisor>=0.0.16,<0.1",
    "jinja2>=3.1.6,<4",
    "twilio>=9.5.1, <10",
    "fastapi>=0.115.12",
    "python-multipart>=0.0.20, <0.1.0",
    "apscheduler>=3.10.1,<4.0.0",
    "dateparser>=1.2.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0"
]

[[project.authors]]
name = "lgesuellip"
email = "lautaro@pampa.ai"

[tool.hatch.build.targets.wheel]
packages = ["src/agents", "src/langgraph_whatsapp"]

[project.scripts]
react-agent = "agents.base.react:main"
